avancement de l'arbre abstrait ajout des constructeurs et override de la fonction operateur ajouts d'un bloc dans l'arbre abstrait à chaque bloc ope_bin retourne la chaine du symbole affectation faites à continuer pour les autres éléments terminaux

+++ ArbreAbstrait/Nombre.java
@@ -8,7 +8,7 @@ public class Nombre extends Expression {
     // Creér à partir de la valeur du nombre
     public Nombre (Integer val, int lig) {
         super(lig);
-        this.valeur = val.intValue();
+        this.valeur = val;
     }
 
     // valeur de l’expression
